2240130-2135 第三次報告版本

+++ src/main/java/com/example/movie/service/impl/BuyInfoServiceImpl.java
@@ -1,35 +1,48 @@
 package com.example.movie.service.impl;
 
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
+import java.security.SecureRandom;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import com.example.movie.constant.RtnCode;
 import com.example.movie.entity.BuyInfo;
 import com.example.movie.repository.BuyInfoDAO;
+import com.example.movie.repository.UserDAO;
 import com.example.movie.service.ifs.BuyInfoService;
 import com.example.movie.vo.BuyInfoGetRes;
 import com.example.movie.vo.UserLoginRes;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 @Service
 public class BuyInfoServiceImpl implements BuyInfoService {
 	
+	private final JavaMailSender javaMailSender;
+	
     @Autowired
     private BuyInfoDAO buyInfoDao;
+    
+    @Autowired
+    private UserDAO userDao;
+    
+    @Autowired
+    public BuyInfoServiceImpl(JavaMailSender javaMailSender) {
+        this.javaMailSender = javaMailSender;
+    }
 
     @Override
     public UserLoginRes create(String account, String movie,int movieId, String cinema, String area, int price,
-			LocalDate onDate, String time, String seat) {
+			LocalDate onDate, String time, String seat,boolean confirmpay) {
         if (!StringUtils.hasText(account)) {
             return new UserLoginRes(RtnCode.ACCOUNT_NOT_FOUND.getCode(),RtnCode.ACCOUNT_NOT_FOUND.getMessage());
         }
@@ -65,15 +78,20 @@ public class BuyInfoServiceImpl implements BuyInfoService {
         		
             }
         }
+        String buyCode = generateVerificationCode();
+        
+        String email = userDao.findUserEmailByAccount(account);
+        
+        sendBuyEmail(email,account,movie,cinema,area,price,onDate,time,seat,buyCode);
 
-        buyInfoDao.save(new BuyInfo(account,movie,movieId,cinema,area,price,onDate,time,seat));
+        buyInfoDao.save(new BuyInfo(account,movie,movieId,cinema,area,price,onDate,time,seat,false));
         
         return new UserLoginRes(RtnCode.SUCCESSFUL.getCode(),RtnCode.SUCCESSFUL.getMessage());
     }
     
 	@Override
 	public UserLoginRes update(int number, String account, String movie,int movieId, String cinema, String area, int price,
-			LocalDate onDate, String time, String seat) {
+			LocalDate onDate, String time, String seat,boolean confirmpay) {
         if (!StringUtils.hasText(seat)) {
             return new UserLoginRes(RtnCode.CHECK_SEAT_INPUT.getCode(),RtnCode.CHECK_SEAT_INPUT.getMessage());
         }
@@ -125,6 +143,9 @@ public class BuyInfoServiceImpl implements BuyInfoService {
         } else {
 			buyinfo.setOnTime(time);
         }
+        if (buyinfo.isConfirmpay() == false) {
+        	buyinfo.setConfirmpay(confirmpay);
+        } 
 		try {
 			buyinfo.setSeat(seat);
 			buyInfoDao.save(buyinfo);
@@ -137,6 +158,11 @@ public class BuyInfoServiceImpl implements BuyInfoService {
 
 	@Override
 	public UserLoginRes delete(int number) {
+		Optional<BuyInfo> op = buyInfoDao.findByNumber(number);
+		BuyInfo buyinfo = op.get();
+        if (buyinfo.isConfirmpay() == true) {
+            return new UserLoginRes(RtnCode.TICKET_IS_PAID.getCode(),RtnCode.TICKET_IS_PAID.getMessage());
+        } 
 		int res = buyInfoDao.deleteByNumber(number);
 		if(res == 0) {
 			return new UserLoginRes(RtnCode.DELETED_BUY_INFO_NOT_EXSISTED.getCode(), RtnCode.DELETED_BUY_INFO_NOT_EXSISTED.getMessage());
@@ -145,6 +171,69 @@ public class BuyInfoServiceImpl implements BuyInfoService {
 		}
 	}
 
+	@Override
+	public UserLoginRes paycheck(int number) {
+        if (number == 0 ) {
+            return new UserLoginRes(RtnCode.ACCOUNT_NOT_FOUND.getCode(),RtnCode.ACCOUNT_NOT_FOUND.getMessage());
+        }
+		Optional<BuyInfo> op = buyInfoDao.findById(number);
+		BuyInfo buyinfo = op.get();
+        if (buyinfo.isConfirmpay() == true ) {
+            return new UserLoginRes(RtnCode.TICKET_IS_PAID.getCode(),RtnCode.TICKET_IS_PAID.getMessage());
+        }
+        
+        String email = userDao.findUserEmailByAccount(buyinfo.getAccount());
+        
+        sendEmail(email,buyinfo.getAccount(),buyinfo.getMovie(),buyinfo.getCinema(),buyinfo.getArea(),buyinfo.getPrice(),buyinfo.getOnDate(),buyinfo.getOnTime(),buyinfo.getSeat());
+        
+        buyinfo.setConfirmpay(true);
+        buyInfoDao.save(buyinfo);
+		return new UserLoginRes(RtnCode.SUCCESSFUL.getCode(), RtnCode.SUCCESSFUL.getMessage());
+	}
+	
+	
+    @Override
+    public void sendEmail(String userEmail,String account, String movie, String cinema, String area, int price,
+			LocalDate onDate, String time, String seat) {
+        try {
+            SimpleMailMessage message = new SimpleMailMessage();
+            message.setFrom("starlightmoviecinema@gmail.com");
+            message.setTo(userEmail);
+            message.setSubject("付費成功通知");
+            
+            // 構建付款頁面URL，將所有訂票相關資訊作為參數添加到URL中
+            String paymentPageUrl = "http://localhost:5173/paypage?" +
+            						"account=" + URLEncoder.encode(account, StandardCharsets.UTF_8) +
+                                    "&movie=" + URLEncoder.encode(movie, StandardCharsets.UTF_8) +
+                                    "&cinema=" + URLEncoder.encode(cinema, StandardCharsets.UTF_8) +
+                                    "&area=" + URLEncoder.encode(area, StandardCharsets.UTF_8) +
+                                    "&price=" + price +
+                                    "&onDate=" + onDate.toString() +
+                                    "&time=" + URLEncoder.encode(time, StandardCharsets.UTF_8) +
+                                    "&seat=" + URLEncoder.encode(seat, StandardCharsets.UTF_8) ;
+
+            // 邮件内容
+            String emailContent = "您的訂票詳情如下：\n" +
+            		"訂購帳號：" + account + "\n" +
+                    "時間：" + onDate.toString() + "，" + time + "\n" +
+                    "電影：" + movie + "\n" +
+                    "影院：" + cinema + "\n" +
+                    "影廳：" + area + "\n" +
+                    "價格：" + price + "元\n" +
+                    "座位：" + seat + "\n\n" +
+                    "付款完畢，祝您觀影愉快！";
+
+            message.setText(emailContent);
+
+            javaMailSender.send(message);
+
+            System.out.println("訂票成功通知郵件已成功發送。");
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+	
+	
 	@Override
 	public UserLoginRes search(String account) {
         if (!StringUtils.hasText(account)) {
@@ -169,6 +258,64 @@ public class BuyInfoServiceImpl implements BuyInfoService {
 		List<BuyInfo> res = new ArrayList<>();
 		res = buyInfoDao.findAllByMovieIdAndCinemaAndAreaAndOnDateAndOnTime(movieId,cinema,area,onDate,time);
 		return new BuyInfoGetRes(RtnCode.SUCCESSFUL.getCode(),RtnCode.SUCCESSFUL.getMessage(),res);
+	}
+	
+	@Override
+	public UserLoginRes searchP(int movieId,String cinema) {
+        if (movieId ==0) {
+            return new UserLoginRes(RtnCode.CHECK_MOVIE_INPUT.getCode(),RtnCode.CHECK_MOVIE_INPUT.getMessage());
+        }
+        Integer res = buyInfoDao.movieprofit(movieId,cinema);
+		return new BuyInfoGetRes(RtnCode.SUCCESSFUL.getCode(),RtnCode.SUCCESSFUL.getMessage(),res);
+	}
+	
+    @Override
+    public void sendBuyEmail(String userEmail,String account, String movie, String cinema, String area, int price,
+			LocalDate onDate, String time, String seat,String buycode) {
+        try {
+            SimpleMailMessage message = new SimpleMailMessage();
+            message.setFrom("starlightmoviecinema@gmail.com");
+            message.setTo(userEmail);
+            message.setSubject("訂票成功通知");
+            
+            // 構建付款頁面URL，將所有訂票相關資訊作為參數添加到URL中
+            String paymentPageUrl = "http://localhost:5173/paypage?" +
+            						"account=" + URLEncoder.encode(account, StandardCharsets.UTF_8) +
+                                    "&movie=" + URLEncoder.encode(movie, StandardCharsets.UTF_8) +
+                                    "&cinema=" + URLEncoder.encode(cinema, StandardCharsets.UTF_8) +
+                                    "&area=" + URLEncoder.encode(area, StandardCharsets.UTF_8) +
+                                    "&price=" + price +
+                                    "&onDate=" + onDate.toString() +
+                                    "&time=" + URLEncoder.encode(time, StandardCharsets.UTF_8) +
+                                    "&seat=" + URLEncoder.encode(seat, StandardCharsets.UTF_8) ;
+
+            // 邮件内容
+            String emailContent = "您的訂票詳情如下：\n" +
+            		"訂購帳號：" + account + "\n" +
+                    "時間：" + onDate.toString() + "，" + time + "\n" +
+                    "電影：" + movie + "\n" +
+                    "影院：" + cinema + "\n" +
+                    "影廳：" + area + "\n" +
+                    "價格：" + price + "元\n" +
+                    "座位：" + seat + "\n\n" +
+                    "您可以點擊下方鏈接進行付款：\n" +
+                    paymentPageUrl + "\n\n" +
+                    "感謝您的訂購，祝您觀影愉快！";
+
+            message.setText(emailContent);
+
+            javaMailSender.send(message);
+
+            System.out.println("訂票成功通知郵件已成功發送。");
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+    
+	private String generateVerificationCode() {
+	    byte[] randomBytes = new byte[16];
+	    new SecureRandom().nextBytes(randomBytes);
+	    return Base64.getEncoder().encodeToString(randomBytes);
 	}
 
 }
